feat: setup CI

+++ tests/run_tests.sh
@@ -1,31 +1,49 @@
 #!/bin/bash
 
+# Only when runing tests on github action CI/CD
+if [ "$1" == "--is_ci" ]; then
+    is_ci=1
+else
+    is_ci=0
+fi
+
 RED="\e[31m"
 GREEN="\e[32m"
 ORANGE="\e[33m"
 END_COLOR="\e[0m"
 
-BUILD_PATH=$PWD/../build
-TESTS_TMP_PATH=$PWD/tmp
+BUILD_PATH=$PWD/build
+TESTS_TMP_PATH=$PWD/tests/tmp
 
-if [ -d ${TESTS_TMP_PATH} ] 
+if [ $is_ci -eq 0 ]
 then
-    rm -rf ${TESTS_TMP_PATH}
+
+    if [ -d ${TESTS_TMP_PATH} ] 
+    then
+        rm -rf ${TESTS_TMP_PATH}
+    fi
+
+    mkdir ${TESTS_TMP_PATH}
+
+    if [ -d ${BUILD_PATH} ] 
+    then
+        rm -rf ${BUILD_PATH}
+    fi
+
+    mkdir ${BUILD_PATH}
+
+    # TODO: add regex as cmd line arg to filter tests per layers (one particuliar test or all test of a specific layer)
+    # TODO: Upload the build directory as artefact and reuse across CI jobs.
+    cd ${BUILD_PATH}
+    cmake .. -DCMAKE_BUILD_TYPE=Release && make -j 16
+    cd ..
+    cp ${BUILD_PATH}/targets/torchinfer ${TESTS_TMP_PATH}
 fi
 
-mkdir ${TESTS_TMP_PATH}
-
-# if [ -d ${BUILD_PATH} ] 
-# then
-#     rm -rf ${BUILD_PATH}
-# fi
-
-# TODO: Upload the build directory as artefact and reuse across CI jobs.
-cmake -B ${BUILD_PATH} .. -DCMAKE_BUILD_TYPE=Release && make -j 16 -C ${BUILD_PATH}
-cp ${BUILD_PATH}/targets/torchinfer ${TESTS_TMP_PATH}
-cd ../
-
-source torchinfer-env/bin/activate
+if [ $is_ci -eq 0 ]
+then
+    source ./torchinfer-env/bin/activate
+fi
 
 mkdir -p ${TESTS_TMP_PATH}/conv2d
 conv2d_list=($(find ${TESTS_TMP_PATH}/../conv2d -name "*.py"))
@@ -40,7 +58,12 @@ do
 
     python ${filename} --output ${TESTS_TMP_PATH}/conv2d
 
-    ${TESTS_TMP_PATH}/torchinfer --input ${TESTS_TMP_PATH}/${testname}_input.bin --type float --onnx_ir ${TESTS_TMP_PATH}/${testname}_ir.bin --output ${TESTS_TMP_PATH}/${testname}_output_cpp.bin
+    if [ $is_ci -eq 0 ]
+    then
+        ${TESTS_TMP_PATH}/torchinfer --input ${TESTS_TMP_PATH}/${testname}_input.bin --type float --onnx_ir ${TESTS_TMP_PATH}/${testname}_ir.bin --output ${TESTS_TMP_PATH}/${testname}_output_cpp.bin
+    else
+        ${BUILD_PATH}/targets/torchinfer --input ${TESTS_TMP_PATH}/${testname}_input.bin --type float --onnx_ir ${TESTS_TMP_PATH}/${testname}_ir.bin --output ${TESTS_TMP_PATH}/${testname}_output_cpp.bin
+    fi
     
     diff ${TESTS_TMP_PATH}/${testname}_output_py.bin ${TESTS_TMP_PATH}/${testname}_output_cpp.bin
